[embedded_cli] avoid conversion warning by changing datatype -> from int to size_t for counter, len, cursor -> propagate size_t change to internal functions

+++ embedded_cli.c
@@ -58,32 +58,32 @@ void embedded_cli_init(struct embedded_cli *cli, const char *prompt,
     embedded_cli_reset_line(cli);
 }
 
-static void cli_ansi(struct embedded_cli *cli, int n, char code)
+static void cli_ansi(struct embedded_cli *cli, size_t n, char code)
 {
     char buffer[5] = {'\x1b', '[', (char) ('0' + (n % 10)), code, '\0'};
     cli_puts(cli, buffer);
 }
 
-static void term_cursor_back(struct embedded_cli *cli, int n)
+static void term_cursor_back(struct embedded_cli *cli, size_t n)
 {
     while (n > 0) {
-        int count = n > 9 ? 9 : n;
+        size_t count = n > 9 ? 9 : n;
         cli_ansi(cli, count, 'D');
         n -= count;
     }
 }
 
-static void term_cursor_fwd(struct embedded_cli *cli, int n)
+static void term_cursor_fwd(struct embedded_cli *cli, size_t n)
 {
     while (n > 0) {
-        int count = n > 9 ? 9 : n;
+        size_t count = n > 9 ? 9 : n;
         cli_ansi(cli, count, 'C');
         n -= count;
     }
 }
 
 #if EMBEDDED_CLI_HISTORY_LEN
-static void term_backspace(struct embedded_cli *cli, int n)
+static void term_backspace(struct embedded_cli *cli, size_t n)
 {
     // printf("backspace %d ('%s': %d)\n", n, cli->buffer, cli->done);
     while (n--)
@@ -111,7 +111,7 @@ static void embedded_cli_insert_default_char(struct embedded_cli *cli,
         return;
     // Insert a gap in the buffer for the new character
     memmove(&cli->buffer[cli->cursor + 1], &cli->buffer[cli->cursor],
-            (size_t) (cli->len - cli->cursor));
+            cli->len - cli->cursor);
     cli->buffer[cli->cursor] = ch;
     cli->len++;
     cli->buffer[cli->len] = '\0';
@@ -184,7 +184,7 @@ static void embedded_cli_stop_search(struct embedded_cli *cli, bool print)
         cli->buffer[sizeof(cli->buffer) - 1] = '\0';
     } else
         cli->buffer[0] = '\0';
-    cli->len = cli->cursor = (int) strlen(cli->buffer);
+    cli->len = cli->cursor = strlen(cli->buffer);
     cli->searching = false;
     if (print) {
         cli_puts(cli, MOVE_BOL CLEAR_EOL);
@@ -205,7 +205,7 @@ bool embedded_cli_insert_char(struct embedded_cli *cli, char ch)
     // printf("Inserting char %d 0x%x '%c'\n", ch, ch, ch);
     if (cli->have_csi) {
         if (ch >= '0' && ch <= '9' && cli->counter < 100) {
-            cli->counter = cli->counter * 10 + ch - '0';
+            cli->counter = (cli->counter * 10) + (size_t) (ch - '0');
             // printf("cli->counter -> %d\n", cli->counter);
         } else {
             if (cli->counter == 0)
@@ -218,7 +218,7 @@ bool embedded_cli_insert_char(struct embedded_cli *cli, char ch)
                 const char *line =
                     embedded_cli_get_history(cli, cli->history_pos + 1);
                 if (line) {
-                    int len = (int) strlen(line);
+                    size_t len = strlen(line);
                     cli->history_pos++;
                     // printf("history up %d = '%s'\n", cli->history_pos,
                     // line);
@@ -246,7 +246,7 @@ bool embedded_cli_insert_char(struct embedded_cli *cli, char ch)
                 const char *line =
                     embedded_cli_get_history(cli, cli->history_pos - 1);
                 if (line) {
-                    int len = (int) strlen(line);
+                    size_t len = strlen(line);
                     cli->history_pos--;
                     // printf("history down %d = '%s'\n",
                     // cli->history_pos, line);
@@ -291,7 +291,7 @@ bool embedded_cli_insert_char(struct embedded_cli *cli, char ch)
                     if (cli->cursor < cli->len) {
                         memmove(&cli->buffer[cli->cursor],
                                 &cli->buffer[cli->cursor + 1],
-                                (size_t) (cli->len - cli->cursor));
+                                cli->len - cli->cursor);
                         cli->len--;
                         cli_puts(cli, &cli->buffer[cli->cursor]);
                         cli_puts(cli, " ");
@@ -346,7 +346,7 @@ bool embedded_cli_insert_char(struct embedded_cli *cli, char ch)
             if (cli->cursor > 0) {
                 memmove(&cli->buffer[cli->cursor - 1],
                         &cli->buffer[cli->cursor],
-                        (size_t) (cli->len - cli->cursor + 1));
+                        cli->len - cli->cursor + 1);
                 cli->cursor--;
                 cli->len--;
                 term_cursor_back(cli, 1);
@@ -374,7 +374,7 @@ bool embedded_cli_insert_char(struct embedded_cli *cli, char ch)
             break;
         case '\x15': // Ctrl-U
             // move back data after cursor, including last \0
-            memmove(cli->buffer, cli->buffer + cli->cursor, (size_t) (cli->len - cli->cursor + 1));
+            memmove(cli->buffer, cli->buffer + cli->cursor, cli->len - cli->cursor + 1);
             cli->len = cli->len - cli->cursor;
             // clear from beggining of buffer,
             // print buffer again and move back to start
